+++ lib/hexapdf/cli.rb
@@ -47,6 +47,7 @@ require 'hexapdf/cli/split'
 require 'hexapdf/cli/watermark'
 require 'hexapdf/cli/image2pdf'
 require 'hexapdf/cli/form'
+require 'hexapdf/cli/fonts'
 require 'hexapdf/version'
 require 'hexapdf/document'
 
@@ -101,6 +102,7 @@ module HexaPDF
         add_command(HexaPDF::CLI::Watermark.new)
         add_command(HexaPDF::CLI::Image2PDF.new)
         add_command(HexaPDF::CLI::Form.new)
+        add_command(HexaPDF::CLI::Fonts.new)
         add_command(CmdParse::HelpCommand.new)
         version_command = CmdParse::VersionCommand.new(add_switches: false)
         add_command(version_command)