+++ lib/hexapdf/cli/images.rb
@@ -197,15 +197,6 @@ module HexaPDF
         end
       end
 
-      # Returns the human readable file size.
-      def human_readable_file_size(size)
-        case size
-        when 0..9999 then "#{size}B"
-        when 10_000..999_999 then "#{(size / 1024.to_f).round(1)}K"
-        else "#{(size.to_f / 1024 / 1024).round(1)}M"
-        end
-      end
-
     end
 
   end